requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data processing (use latest compatible versions)
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Machine Learning and AI
openai>=1.0.0
anthropic>=0.7.0

# Data validation and models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# File handling and security
werkzeug>=3.0.0
python-multipart>=0.0.6

# Streamlit for UI
streamlit>=1.28.0
plotly>=5.15.0
requests>=2.28.0

# Environment management
python-dotenv>=1.0.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0

# Additional utilities
click>=8.1.0
rich>=13.0.0